Move filter_events_for_server out of FederationHandler
for easier unit testing.
